# docs/swagger.yaml
basePath: /api
definitions:
...
  database.Endpoint:
    properties:
      agent_version:
        $ref: '#/definitions/pgtype.Text'
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      deployment_id:
        type: string
      first_seen:
        $ref: '#/definitions/pgtype.Timestamptz'
      hostname:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      last_seen:
        $ref: '#/definitions/pgtype.Timestamptz'
      os:
        $ref: '#/definitions/pgtype.Text'
      status:
        $ref: '#/definitions/database.EndpointStatus'
      tenant_id:
        type: string
    type: object
  database.EndpointStatus:
    enum:
    - active
    - inactive
    - deleted
    type: string
    x-enum-varnames:
    - EndpointStatusActive
    - EndpointStatusInactive
    - EndpointStatusDeleted
  database.EntitlementEvent:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      deployment_id:
        type: string
      event_type:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      subscription_id:
        type: string
      tenant_id:
        type: string
    type: object
...
  handlers.ActiveEndpointCountResponse:
    properties:
      count:
        example: 42
        type: integer
    type: object
...
  handlers.CreateEntitlementEventRequest:
    properties:
      deployment_id:
        example: 123e4567-e89b-12d3-a456-426614174002
        type: string
      event_type:
        example: feature_enabled
        type: string
      metadata:
        additionalProperties: {}
        type: object
      subscription_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.RegisterEndpointRequest:
    properties:
      agent_version:
        example: 2.1.0
        type: string
      deployment_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      hostname:
        example: agent-prod-01.acme.com
        type: string
      os:
        example: linux
        type: string
    type: object
  handlers.SearchEventsRequest:
    properties:
      limit:
        example: 50
        type: integer
      metadata:
        additionalProperties: {}
        type: object
    type: object
  handlers.UpdateEndpointStatusRequest:
    properties:
      status:
        example: inactive
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: Cloud License Platform — multi-tenant SaaS licensing and entitlement
    management.
  title: CLP API
  version: "1.0"
paths:
...
  /endpoints/{id}:
    delete:
      description: Permanently removes an endpoint/agent by UUID.
      parameters:
      - description: Endpoint UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete an endpoint
      tags:
      - endpoints
    get:
      description: Returns an endpoint/agent by UUID.
      parameters:
      - description: Endpoint UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Endpoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get an endpoint
      tags:
      - endpoints
  /endpoints/{id}/heartbeat:
    post:
      description: Updates the last_seen timestamp for the endpoint to now.
      parameters:
      - description: Endpoint UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Record endpoint heartbeat
      tags:
      - endpoints
  /endpoints/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets the endpoint status (active, inactive, deleted).
      parameters:
      - description: Endpoint UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateEndpointStatusRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update endpoint status
      tags:
      - endpoints
  /events:
    get:
      description: Returns entitlement events filtered by event type, newest first.
      parameters:
      - description: Event type
        in: query
        name: type
        required: true
        type: string
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.EntitlementEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List events by type
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Records an entitlement lifecycle event with arbitrary metadata.
      parameters:
      - description: Event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEntitlementEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.EntitlementEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create an entitlement event
      tags:
      - events
  /events/search:
    post:
      consumes:
      - application/json
      description: Returns events whose metadata contains the provided JSON fragment
        (JSONB containment query).
      parameters:
      - description: Search query
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SearchEventsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.EntitlementEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Search events by metadata
      tags:
      - events
...
  /tenants/{id}/endpoints:
    get:
      description: Returns all endpoints/agents registered under the given tenant,
        newest first.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Endpoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List endpoints for a tenant
      tags:
      - endpoints
    post:
      consumes:
      - application/json
      description: Registers a new agent/endpoint under the given tenant.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Endpoint
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterEndpointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Endpoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register an endpoint for a tenant
      tags:
      - endpoints
  /tenants/{id}/endpoints/count:
    get:
      description: Returns the number of endpoints with status "active" for the given
        tenant.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ActiveEndpointCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Count active endpoints for a tenant
      tags:
      - endpoints
  /tenants/{id}/events:
    get:
      description: Returns entitlement events for the given tenant, newest first,
        with pagination.
      parameters:
      - description: Tenant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Page size (default 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.EntitlementEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List events for a tenant
      tags:
      - events
...
swagger: "2.0"